% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_revenue_regression.R
\name{compute_revenue_regression}
\alias{compute_revenue_regression}
\title{How does weather influence baseball ticket sales?}
\usage{
compute_revenue_regression(revenue, storm, temp, humid)
}
\arguments{
\item{revenue}{vector containing ticket sales revenue for all home games}

\item{storm}{vector containing information on whether or not there was a thunderstorm the day of the baseball game}

\item{temp}{vector containing the mean temperature for the day the baseball game was on}

\item{humid}{vector containing the max humidity for the day the baseball game was on}
}
\value{
a summary output of the regression
}
\description{
Computes whether there is a noticable association between weather and average ticket sales revenue
}
