% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_logistic.R
\name{homerun_probability_computation}
\alias{homerun_probability_computation}
\title{How does weather influence the probability of hitting a homerun?}
\usage{
homerun_probability_computation(homerun_array, temp_array, humid_array,
  temperature, humidity)
}
\arguments{
\item{humid_array}{vector containing the max humidity for the day the baseball game was on}

\item{temperature}{a temperature of interest to plug into the binomial logistic regression}

\item{humidity}{a humidity of interest to plug into the binomial logistic regression}

\item{homerun}{Binoary response variable stating whether or not a home run was hit during that days game}

\item{temp_arrau}{vector containing the mean temperature for the day the baseball game was on}
}
\value{
The probability of there being a home run based on your inputted arrays and temperature and humidity of interest
}
\description{
Computes whether there is a noticable association between weather and average ticket sales revenue
}
