% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeruns.R
\name{homeruns}
\alias{homeruns}
\title{Homerun Probability Increases From Humidity}
\usage{
homeruns(ah)
}
\arguments{
\item{ah}{is the output from the SeasonHumd function
This assumes that humidity below 50 percent will not have a significant effect on homeruns}
}
\description{
This function converts the average humidity to increases in homerun probabilities.
}
