% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/june.R
\name{june}
\alias{june}
\title{Season Temperature Calculator}
\usage{
june(city)
}
\arguments{
\item{city}{is your input table. See the package description and example data to see the format required(View(bronx)).}
}
\description{
This function will return a tibble with the average high temperature for each of the months of your baseball season
}
