% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonHumd.R
\name{SeasonHumd}
\alias{SeasonHumd}
\title{Humidity Summary}
\usage{
SeasonHumd(city)
}
\arguments{
\item{city}{is your input table. See the package description and example data to see the format required
This function will return a tibble with the average humidity for each of the months of your baseball season}
}
\description{
Humidity Summary
}
