% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_temperature_computation.R
\name{yearly_temperature_computation}
\alias{yearly_temperature_computation}
\title{Compute yearly average across different cities that have baseball teams.}
\usage{
yearly_temperature_computation(temperature_array)
}
\arguments{
\item{temperature_array}{a multidimensional array containing monthly temperatures across years for different cities}
}
\description{
This function will return the yearly average temperature across multiple cities.
}
