% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeruns.R
\name{homeruns}
\alias{homeruns}
\title{homeruns summary}
\usage{
homeruns(ah)
}
\arguments{
\item{ah}{is the output from the SeasonHumd function
This assumes that humidity below 50% will not have a significant effect on homeruns}
}
\description{
homeruns summary
}
