% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/june.R
\name{june}
\alias{june}
\title{June}
\usage{
june(city)
}
\arguments{
\item{city}{is your input table. See the package description and example data to see the format required
'This function will return a tibble with the average humidity for each of the months of your baseball season}
}
\description{
June
}
